\name{rescaleP}
\alias{rescaleP}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Rescale \code{pvec} in function \code{Bisect}
}
\description{
An internal function  borrowed from \code{Icens}.
}
%\usage{
%}
%- maybe also 'usage' for other objects documented here.
%\arguments{
%}
%\details{
%}
%\value{
%}
%\references{
%}
\author{
Alain Vandal and Robert Gentleman
}
\note{ 
It is not to be called by the user.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
 \code{\link{Bisect}}
}
%\examples 
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
