\name{Teststat4}
\alias{Teststat4}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Compute the test statistic in test gLRT4
}
\description{
The function computes the test statistics of the test proposed by Zhao, Duan, Zhao, and Sun (2013).
}

\references{
X. Zhao, R. Duan, Q. Zhao, and J. Sun (2013), "A New Class of Generalized Log Rank Tests for Interval-censored Failure Time Data", Computational Statistics and Data Analysis. 60: 123-131.
}

\author{
Qiang Zhao and Jianguo Sun
}

\note{
It is not to be called by the user.
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{gLRT4}}
}
%\examples{
%}
